+++ src/pages/_app.tsx
@@ -3,7 +3,7 @@ import { QueryClientProvider } from "@tanstack/react-query";
 import { ReactQueryDevtools } from "@tanstack/react-query-devtools";
 import type { AppProps } from "next/app";
 import { queryClient } from "../setup";
-import { useEffect } from "react";
+import { useEffect, useLayoutEffect } from "react";
 
 export default function App({ Component, pageProps }: AppProps) {
   // useEffect(() => {
